Fix all known problems with relative pathnames.
Git's --work-tree option made the names relative
to the wrong location, and was unnecessary for fetch
anyway; remove.

Git's file:// syntax cannot cope with relative paths
with slashes in them, apparently due to some problem
with parsing the first segment as a "hostname". Not
using the file:// syntax means paths with colons are
mistaken for ssh URLs; I have no workaround for this.